+++ src/Controllers/SiteController.php
@@ -184,7 +184,6 @@ class SiteController extends Controller
         return Entry::find()
             ->selectSiteAttributes()
             ->replaceI18nAttributes()
-            ->addSelectI18nSlugTargetAttributes()
             ->whereStatus($status)
             ->andWhereParentStatus();
     }